detile: major update to detiler
Update the detiler to accept command line options for width and height,
as well as selecting between normal tiling and multitiling modes.

Optionally accept the file to be detiled on the command line too, and
read it into memory if we aren't able to mmap() it.

Lastly, improve the safety of writing the resulting image out by
handling short writes and signals gracefully.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>